added files to component checkout
