+++ client/src/components/Checkout/pages/BagPage/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './BagPage';
