+++ client/src/components/Checkout/pages/EmptyListPage/EmptyList.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import styles from './EmptyList.module.scss';
+import Button from '../../../Button';
+import TabsSection from '../../sections/TabsSection';
+import Header from '../../../Header';
+const EmptyList = () => {
+    return (
+        <>
+            <Header />
+            <TabsSection />
+            <div className={styles.emptyMain}>
+                <h1 className={styles.emptyTitle}>THANK YOU FOR YOUR ORDER!</h1>
+                <Button className={styles.emptyButton} text={'CONTINUE SHOPPING'} />
+            </div>
+        </>
+    );
+};
+export default EmptyList;
