+++ client/src/components/Checkout/pages/EmptyListPage/EmptyList.module.scss
@@ -1,4 +1,4 @@
-@import '../../../style/mixins';
+@import '../../../../style/mixins';
 .emptyMain {
 
 }
