+++ client/src/components/Checkout/pages/EmptyListPage/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './EmptyList';
