+++ client/src/components/Checkout/pages/PaymentOptionsPage/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './PaymentOptionsPage';
