+++ client/src/components/Checkout/pages/ShippingDetailsPage/ShippingDetailsPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../../../Header';
+import TabsSection from '../../sections/TabsSection';
+const ShippingDetailsPage = () => {
+    return (
+        <>
+            <div>
+                <Header />
+                <TabsSection />
+            </div>
+        </>
+    );
+};
+export default ShippingDetailsPage;
