+++ client/src/components/Checkout/pages/ShippingDetailsPage/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './ShippingDetailsPage';
