+++ client/src/components/Checkout/sections/PaymentSection/PaymentSection.module.scss
@@ -1,4 +1,4 @@
-@import '../../../style/mixins';
+@import '../../../../style/mixins';
 
 .container {
   @include content-container;
