+++ client/src/components/Checkout/sections/PaymentSection/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './PaymentSection';
