+++ client/src/components/Checkout/sections/SummarySection/SummarySection.module.scss
@@ -1,4 +1,4 @@
-@import '../../../style/mixins';
+@import '../../../../style/mixins';
 
 .container {
   @include content-container;
