+++ client/src/components/Checkout/sections/SummarySection/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SummarySection';
