+++ client/src/components/Checkout/sections/TabsSection/TabsSection.module.scss
@@ -1,9 +1,9 @@
-@import '../../../style/mixins';
+@import '../../../../style/mixins';
 .tabsMain {
   width: 100%;
   height: 10rem;
   background: #373F41;
-  margin-top: 20rem;
+  margin-top: -3rem;
 }
 .tabsNav {
   @include content-container;
