+++ client/src/components/Checkout/sections/TabsSection/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './TabsSection';
